requirements.txt
Flask==3.0.0
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.25.2
